+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog].
 * Use `perltidy` as default formatter for `cperl-mode` ([#260]).
 * The `perltidy` formatter now supports Emacs indentation and line length
   settings ([#261]).
+* Indent level support for yq JSON, XML, and YAML ([#259]).
 
 ### Formatters
 * [`js-beautify`](https://github.com/beautify-web/js-beautify) for
@@ -22,6 +23,7 @@ The format is based on [Keep a Changelog].
   js, jsx, ts, tsx, json, jsonc, md files. ([#264])
 
 [#229]: https://github.com/radian-software/apheleia/pull/229
+[#259]: https://github.com/radian-software/apheleia/pull/259
 [#260]: https://github.com/radian-software/apheleia/pull/260
 [#261]: https://github.com/radian-software/apheleia/pull/261
 [#263]: https://github.com/radian-software/apheleia/pull/263